Add abstract Enitity parent class for Feed and Item
